// src/games/whack/art.generated.ts
export const KEY_POSES = {
  park: {
    glance: require('../../assets/games/whack/themes/park/glance.png'),
    contact: require('../../assets/games/whack/themes/park/contact.png'),
    bonked_spiral: require('../../assets/games/whack/themes/park/bonked_spiral.png'),
    bonked_tongue: require('../../assets/games/whack/themes/park/bonked_tongue.png'),
    bonked_hatoff: require('../../assets/games/whack/themes/park/bonked_hatoff.png'),
    duck: require('../../assets/games/whack/themes/park/duck.png'),
  },
  pirates: {
    glance: require('../../assets/games/whack/themes/pirates/glance.png'),
    contact: require('../../assets/games/whack/themes/pirates/contact.png'),
    bonked_spiral: require('../../assets/games/whack/themes/pirates/bonked_spiral.png'),
    bonked_tongue: require('../../assets/games/whack/themes/pirates/bonked_tongue.png'),
    bonked_hatoff: require('../../assets/games/whack/themes/pirates/bonked_hatoff.png'),
    duck: require('../../assets/games/whack/themes/pirates/duck.png'),
  },
  mansion: {
    glance: require('../../assets/games/whack/themes/mansion/glance.png'),
    contact: require('../../assets/games/whack/themes/mansion/contact.png'),
    bonked_spiral: require('../../assets/games/whack/themes/mansion/bonked_spiral.png'),
    bonked_tongue: require('../../assets/games/whack/themes/mansion/bonked_tongue.png'),
    bonked_hatoff: require('../../assets/games/whack/themes/mansion/bonked_hatoff.png'),
    duck: require('../../assets/games/whack/themes/mansion/duck.png'),
  },
  space: {
    glance: require('../../assets/games/whack/themes/space/glance.png'),
    contact: require('../../assets/games/whack/themes/space/contact.png'),
    bonked_spiral: require('../../assets/games/whack/themes/space/bonked_spiral.png'),
    bonked_tongue: require('../../assets/games/whack/themes/space/bonked_tongue.png'),
    bonked_hatoff: require('../../assets/games/whack/themes/space/bonked_hatoff.png'),
    duck: require('../../assets/games/whack/themes/space/duck.png'),
  },
  jungle: {
    glance: require('../../assets/games/whack/themes/jungle/glance.png'),
    contact: require('../../assets/games/whack/themes/jungle/contact.png'),
    bonked_spiral: require('../../assets/games/whack/themes/jungle/bonked_spiral.png'),
    bonked_tongue: require('../../assets/games/whack/themes/jungle/bonked_tongue.png'),
    bonked_hatoff: require('../../assets/games/whack/themes/jungle/bonked_hatoff.png'),
    duck: require('../../assets/games/whack/themes/jungle/duck.png'),
  },
  backlot: {
    glance: require('../../assets/games/whack/themes/backlot/glance.png'),
    contact: require('../../assets/games/whack/themes/backlot/contact.png'),
    bonked_spiral: require('../../assets/games/whack/themes/backlot/bonked_spiral.png'),
    bonked_tongue: require('../../assets/games/whack/themes/backlot/bonked_tongue.png'),
    bonked_hatoff: require('../../assets/games/whack/themes/backlot/bonked_hatoff.png'),
    duck: require('../../assets/games/whack/themes/backlot/duck.png'),
  },
} as const;
// ...
export const HATS: Record<WhackThemeId, number> = {
  park: require('../../assets/games/whack/themes/park/hat.png'),
  pirates: require('../../assets/games/whack/themes/pirates/hat.png'),
  mansion: require('../../assets/games/whack/themes/mansion/hat.png'),
  space: require('../../assets/games/whack/themes/space/hat.png'),
  jungle: require('../../assets/games/whack/themes/jungle/hat.png'),
  backlot: require('../../assets/games/whack/themes/backlot/hat.png'),
};
export const HAT_BOX: Record<WhackThemeId, Box> = {
  park: [0.038, 0.057, 0.965, 0.948, 1.485],
  pirates: [0.038, 0.06, 0.965, 0.94, 1.574],
  mansion: [0.04, 0.038, 0.96, 0.962, 0.962],
  space: [0.035, 0.038, 0.962, 0.958, 1.108],
  jungle: [0.038, 0.053, 0.962, 0.947, 1.391],
  backlot: [0.038, 0.057, 0.965, 0.948, 1.492],
};

export const DECKS: Record<WhackThemeId, number> = {
  park: require('../../assets/games/whack/decks/park.jpg'),
  pirates: require('../../assets/games/whack/decks/pirates.jpg'),
  mansion: require('../../assets/games/whack/decks/mansion.jpg'),
  space: require('../../assets/games/whack/decks/space.jpg'),
  jungle: require('../../assets/games/whack/decks/jungle.jpg'),
  backlot: require('../../assets/games/whack/decks/backlot.jpg'),
};

export const RIM_WARNING = { teeth: require('../../assets/games/whack/rims/warning_teeth.png'), tab: require('../../assets/games/whack/rims/warning_tab.png') };
export const RIM_WARNING_BOX = { teeth: [0.038, 0.056, 0.962, 0.944, 1.609] as Box, tab: [0.044, 0.035, 0.956, 0.965, 0.788] as Box };
export const SPEEDLINES = require('../../assets/games/whack/fx/speedlines.png');

export const PROPS = {
  pirates_bow: require('../../assets/games/whack/props/pirates_bow.png'),
  mansion_gate: require('../../assets/games/whack/props/mansion_gate.png'),
  space_gantry: require('../../assets/games/whack/props/space_gantry.png'),
  jungle_steps: require('../../assets/games/whack/props/jungle_steps.png'),
  backlot_light_a: require('../../assets/games/whack/props/backlot_light_a.png'),
  backlot_light_b: require('../../assets/games/whack/props/backlot_light_b.png'),
  backlot_chair: require('../../assets/games/whack/props/backlot_chair.png'),
};